requirements.txt
anthropic
requests
yfinance
pandas
python-dotenv